Engine: fix value transaltion.
In the call back of `SBUFetch` when asking the ledger for a contract the translateValue may raise an exception.
However it is not safe as the engine cannot recover such error.

We create a new Control for the speedy machine, that handle safely the translation.

CHANGELOG_BEGIN
CANGELOG_END
