+++ daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/SBuiltin.scala
@@ -13,7 +13,13 @@ import com.daml.lf.data.Numeric.Scale
 import com.daml.lf.language.Ast
 import com.daml.lf.speedy.SError._
 import com.daml.lf.speedy.SExpr._
-import com.daml.lf.speedy.Speedy.{CtrlValue, CtrlWronglyTypeContractId, Machine, SpeedyHungry}
+import com.daml.lf.speedy.Speedy.{
+  CtrlTranslateValue,
+  CtrlValue,
+  CtrlWronglyTypeContractId,
+  Machine,
+  SpeedyHungry
+}
 import com.daml.lf.speedy.SResult._
 import com.daml.lf.speedy.SValue._
 import com.daml.lf.transaction.{Transaction => Tx}
@@ -905,101 +911,6 @@ object SBuiltin {
     }
   }
 
-  // This function translates well-typed values.
-  // Raises an exception if missing packages.
-  @throws[SpeedyHungry]
-  def translateValue(machine: Machine, value: V[V.ContractId]): CtrlValue = {
-    def go(value0: V[V.ContractId]): SValue =
-      value0 match {
-        case V.ValueList(vs) => SList(vs.map[SValue](go))
-        case V.ValueContractId(coid) => SContractId(coid)
-        case V.ValueInt64(x) => SInt64(x)
-        case V.ValueNumeric(x) => SNumeric(x)
-        case V.ValueText(t) => SText(t)
-        case V.ValueTimestamp(t) => STimestamp(t)
-        case V.ValueParty(p) => SParty(p)
-        case V.ValueBool(b) => SBool(b)
-        case V.ValueDate(x) => SDate(x)
-        case V.ValueUnit => SUnit
-        case V.ValueRecord(Some(id), fs) =>
-          val fields = Name.Array.ofDim(fs.length)
-          val values = new util.ArrayList[SValue](fields.length)
-          fs.foreach {
-            case (optk, v) =>
-              optk match {
-                case None =>
-                  throw crash("SValue.fromValue: record missing field name")
-                case Some(k) =>
-                  fields(values.size) = k
-                  val _ = values.add(go(v))
-              }
-          }
-          SRecord(id, fields, values)
-        case V.ValueRecord(None, _) =>
-          crash("SValue.fromValue: record missing identifier")
-        case V.ValueStruct(fs) =>
-          val fields = Name.Array.ofDim(fs.length)
-          val values = new util.ArrayList[SValue](fields.length)
-          fs.foreach {
-            case (k, v) =>
-              fields(values.size) = k
-              val _ = values.add(go(v))
-          }
-          SStruct(fields, values)
-        case V.ValueVariant(None, _variant @ _, _value @ _) =>
-          crash("SValue.fromValue: variant without identifier")
-        case V.ValueEnum(None, constructor @ _) =>
-          crash("SValue.fromValue: enum without identifier")
-        case V.ValueOptional(mbV) =>
-          SOptional(mbV.map(go))
-        case V.ValueTextMap(map) =>
-          STextMap(map.mapValue(go).toHashMap)
-        case V.ValueGenMap(entries) =>
-          SGenMap(
-            entries.iterator.map { case (k, v) => go(k) -> go(v) }
-          )
-        case V.ValueVariant(Some(id), variant, value) =>
-          machine.compiledPackages.getPackage(id.packageId) match {
-            case Some(pkg) =>
-              pkg.lookupIdentifier(id.qualifiedName).fold(crash, identity) match {
-                case Ast.DDataType(_, _, data: Ast.DataVariant) =>
-                  SVariant(id, variant, data.constructorRank(variant), go(value))
-                case _ =>
-                  crash(s"definition for variant $id not found")
-              }
-            case None =>
-              throw SpeedyHungry(
-                SResultNeedPackage(
-                  id.packageId,
-                  pkg => {
-                    machine.compiledPackages = pkg
-                    machine.ctrl = translateValue(machine, value)
-                  }
-                ))
-          }
-        case V.ValueEnum(Some(id), constructor) =>
-          machine.compiledPackages.getPackage(id.packageId) match {
-            case Some(pkg) =>
-              pkg.lookupIdentifier(id.qualifiedName).fold(crash, identity) match {
-                case Ast.DDataType(_, _, data: Ast.DataEnum) =>
-                  SEnum(id, constructor, data.constructorRank(constructor))
-                case _ =>
-                  crash(s"definition for variant $id not found")
-              }
-            case None =>
-              throw SpeedyHungry(
-                SResultNeedPackage(
-                  id.packageId,
-                  pkg => {
-                    machine.compiledPackages = pkg
-                    machine.ctrl = translateValue(machine, value)
-                  }
-                ))
-          }
-      }
-    CtrlValue(go(value))
-  }
-
   /** $fetch[T]
     *    :: ContractId a
     *    -> Token
@@ -1039,7 +950,7 @@ object SBuiltin {
                             CtrlWronglyTypeContractId(acoid, templateId, coinst.template)
                         } else {
                           machine.ptx = machine.ptx.cachedContract(coid, coinst)
-                          machine.ctrl = translateValue(machine, coinst.arg.value)
+                          machine.ctrl = CtrlTranslateValue(coinst.arg.value)
                         }
                     },
                   ),
@@ -1060,7 +971,7 @@ object SBuiltin {
         // (see below).
         crash(s"Relative contract $coid ($templateId) not found from partial transaction")
       } else {
-        machine.ctrl = translateValue(machine, coinst.arg.value)
+        machine.ctrl = CtrlTranslateValue(coinst.arg.value)
       }
     }
   }
